backend/requirements.txt
# Framework
fastapi==0.111.0
uvicorn==0.30.0

# Data
requests>=2.31.0
pandas>=2.2.2
ta>=0.11.0
numpy>=1.26.4

# IA
google-generativeai==0.7.0

# Banco de dados
supabase==2.5.0

# Utilitários
python-dotenv==1.0.1
tenacity==8.3.0
pydantic==2.7.0
pydantic-settings==2.3.0
httpx==0.27.0

# Testes
pytest==8.2.0
pytest-asyncio==0.23.7
